UrlSubmitter: Set error code and text
When there is an error, set the response code and text so that the user is notified in an alert.
Set code to 550 which is not a designated HTTP error code as we don't know what the error is. User will be prompted to look at the browser console as with other non-200 responses.